requirements/requirements.txt
asgiref==3.4.1
astral==2.2
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
derek-functions @ file:///home/derek/derekModule
Django==4.0.1
djangorestframework==3.13.1
idna==3.2
mysql-connector==2.2.9
mysqlclient==2.1.0
paramiko==2.7.2
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.4.0
pysftp==0.2.9
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
twilio==6.63.1
urllib3==1.26.6
